Changed the radius to an unsigned int rather than NSUInteger to squelch the Xcode warning.

+++ Onboard/OnboardingViewController.m
@@ -252,7 +252,7 @@ static CGFloat const kDefaultSaturationDeltaFactor = 1.8;
             // ... if d is odd, use three box-blurs of size 'd', centered on the output pixel.
             //
             CGFloat inputRadius = blurRadius * [[UIScreen mainScreen] scale];
-            NSUInteger radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
+            unsigned int radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
             if (radius % 2 != 1) {
                 radius += 1; // force radius to be odd so that the three box-blur methodology works.
             }
